requirements.txt
streamlit
pymongo
faiss-cpu
sentence-transformers
spacy
pandas
numpy
scikit-learn
